integration des fichiers miniatures dans les fichiers videos pour resoudre les conflits, correction des noms de routes
