+++ backend/src/models/VideoManager.js
@@ -23,5 +23,24 @@ class MainVideoPlayerManager extends AbstractManager {
     );
     return rows;
   }
+
+  async readByCategories(category, limit) {
+    let sql = `SELECT title, image FROM video
+    inner join video_category on video_category.video_id = video.id
+    inner join category on category.id = video_category.category_id `;
+    const sqlValues = [];
+    if (category != null) {
+      sql += "where category.name = ?";
+      sqlValues.push(category);
+    }
+    if (limit != null) {
+      sql += " limit ?";
+      sqlValues.push(limit);
+    }
+
+    const [rows] = await this.database.query(sql, sqlValues);
+
+    return rows;
+  }
 }
 module.exports = MainVideoPlayerManager;
