+++ backend/src/tables.js
@@ -5,12 +5,10 @@
 // Import the manager modules responsible for handling data operations on the tables
 const ItemManager = require("./models/ItemManager");
 const VideoManager = require("./models/VideoManager");
-const MiniatureManager = require("./models/MiniatureManager");
 
 const managers = [
   ItemManager,
   VideoManager,
-  MiniatureManager,
   // Add other managers here
 ];
