+++ frontend/src/components/boutonsLanguages/BoutonsLanguages.jsx
@@ -12,7 +12,7 @@ function VideoPage() {
   useEffect(() => {
     (async () => {
       const triedCategory = await fetch(
-        `http://localhost:3310/api/miniatures?category=${actualCategory}`
+        `http://localhost:3310/api/videosSelected?category=${actualCategory}`
       );
       await triedCategory.json();
     })();
